refactor: Reorganize application structure and remove redundant app package
- Remove internal/app package and distribute its functionality
- Move SearchContent function to new internal/search package
- Simplify NewCrawler function by moving it directly into crawler module
- Update import paths in cmd/search.go and cmd/crawl.go
- Remove unnecessary app-related modules and files
